+++ dev/stories/0006-hapi-startup-wait.md
@@ -1,0 +1,10 @@
+# Bug: HAPI Failures After Launch
+
+Intermittently, HAPI is experiencing test case failures early on in a run.
+These failures look like this in the log:
+
+```
+TODO: get these from a Linux or WSL 2 box
+```
+
+These errors _seem_ to clear up once HAPI has been running for a little bit.
