+++ fhir-bench-orchestrator/Cargo.toml
@@ -36,3 +36,6 @@ rust_decimal = { version = "1", features = ["serde-float"] }
 
 # Date & time.
 chrono = { version = "0.4", features = ["serde"] }
+
+# Used in integration tests, to run binaries and verify results.
+assert_cmd = "1"
