+++ fhir-bench-orchestrator/src/servers/hapi_jpa.rs
@@ -4,22 +4,25 @@ use anyhow::{anyhow,Context,Result};
 use std::process::Command;
 use url::Url;
 
-static SERVER_NAME_TEXT: &str = "HAPI FHIR JPA Server";
-static SERVER_NAME: ServerName = ServerName(SERVER_NAME_TEXT);
+static SERVER_NAME: &str = "HAPI FHIR JPA Server";
 
 /// The trait object for the `ServerPlugin` implementation for the HAPI FHIR JPA server.
-pub struct HapiJpaFhirServerPlugin {}
+pub struct HapiJpaFhirServerPlugin {
+    server_name: ServerName
+}
 
 impl HapiJpaFhirServerPlugin {
     /// Constructs a new `HapiJpaFhirServerPlugin` instance.
     pub fn new() -> HapiJpaFhirServerPlugin {
-        HapiJpaFhirServerPlugin {}
+        HapiJpaFhirServerPlugin {
+            server_name: SERVER_NAME.into()
+        }
     }
 }
 
 impl ServerPlugin for HapiJpaFhirServerPlugin {
-    fn server_name(&self) -> &'static ServerName {
-        &SERVER_NAME
+    fn server_name(&self) -> &ServerName {
+        &self.server_name
     }
 
     fn launch(&self) -> Result<Box<dyn ServerHandle>> {
