+++ fhir-bench-orchestrator/src/test_framework/metadata.rs
@@ -2,18 +2,16 @@
 
 use crate::AppState;
 use crate::servers::ServerHandle;
-use crate::test_framework::{ServerOperationLog, ServerOperationName};
+use crate::test_framework::ServerOperationLog;
 use chrono::prelude::*;
 use slog::warn;
 
-static SERVER_OP_NAME_METADATA_TEXT: &str = "metadata";
-static SERVER_OP_NAME_METADATA: ServerOperationName =
-    ServerOperationName(SERVER_OP_NAME_METADATA_TEXT);
+static SERVER_OP_NAME_METADATA: &str = "metadata";
 
 /// Verifies and benchmarks FHIR `/metadata` operations.
 pub fn run_operation_metadata(app_state: &AppState, server_handle: &dyn ServerHandle) -> ServerOperationLog {
     let mut server_op_log = ServerOperationLog {
-        operation: &SERVER_OP_NAME_METADATA,
+        operation: SERVER_OP_NAME_METADATA.into(),
         started: Utc::now(),
         iterations: app_state.config.iterations,
         completed: None,
